Tenerife Island Volcanic Risk Action Plan (PAIV)
The Tenerife Island Volcanic Risk Action Plan (PAIV for its abbreviation in Spanish) is the technical instrument at the island level that establishes the organizational and functional structure of all resources and personnel involved during an emergency due to volcanic risk on the island. It also defines the operational procedures and coordination mechanisms, while fully respecting its organizational autonomy, to carry out the functions assigned to it by the Special Plan of the Autonomous Community.